dbquery final version
Working version of dbquery that can scan the whole heap file to find matches. The previous iteration had a problem where it skipped the beginning of a page, resulting in incorrect ordering and wrong value of the records. This version fixes that problem, along with fixing the problem where the program returned EndOfFile Exception at the end of the heap file. The solution was by stopping the loop when the maximum number of pages is reached. This version adds total operation time in the result file as well.